cpp01 update

+++ 01/ex04/src/main.cpp
@@ -6,7 +6,7 @@
 /*   By: sanan <sanan@student.42seoul.kr>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/22 22:34:18 by sanan             #+#    #+#             */
-/*   Updated: 2023/03/23 18:46:47 by sanan            ###   ########.fr       */
+/*   Updated: 2023/04/12 15:06:37 by sanan            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,15 +14,14 @@
 #include <iostream>
 
 std::string replaceString(std::string origin, std::string from, std::string to) {
-    std::string result;
-    size_t start_pos = 0;
-    size_t pos;
+    std::string		result;
+    unsigned int	start_pos = 0;
+    unsigned int	pos;
 
     while ((pos = origin.find(from, start_pos)) != std::string::npos) {
         result += origin.substr(start_pos, pos - start_pos) + to;
         start_pos = pos + from.length();
     }
-
     result += origin.substr(start_pos);
 	return (result);
 }
